Add PMOVMSKB to x86 model
The manual is a bit confusing about the destination size, and
assemblers differ in whether they allow 64-bit destination
registers. But the preferred disassembly seems to be 32-bit.
Since the result is just a zero-extension of a 16-bit value,
and 32-bit writes are then zero-extended to 64 bits anyway,
it makes no semantic difference.