AVX2 instructions needed for ML-KEM code (#291)
* Add VMOVDQU to x86 model

This also fixes an error in read_VEX, which was not correctly
analyzing the implicit prefix parts in the case of a 2-byte
VEX prefix, only doing so in the 3-byte case.

* Add VPOR to the x86 model

* Add VPSLLW to x86 model

* Add VPSLLD and VPSRLD (usual immediate third operand) to x86 model

* Add VPBLENDW to the x86 model

This also enforces the ".W0" (VEX.W = 0) restriction on the existing
decoding of VPBLENDD. For the new VPBLENDW by contrast, the field is
ignored so there are acceptable encodings with an extra bit set.

* Add VPERMD to x86 model

* Add VPSHUFB to x86 model

* Tighten up some W0 and W1 restrictions on AVX2 instructions

Many of these are to be decoded as "WIG", i.e. the W field ignored,
but a few require a specific setting 0 or 1 for the W field according
to the manual. This makes sure we don't accept other settings, which
are probably illegal or distinct instructions.

* Refine implicit prefix requirements on AVX2 instructions

This tightens up the decoding to require the actual implicit
prefixes as specified in the manual by pattern-matching over
the "pfxs" field extracted by read_VEXP. The correspondence
with the notation in the manual is:

    (T, Rep0, SG0) corresponds to VEX.nnn.66
    (F, RepZ, SG0) corresponds to VEX.nnn.f3
    (F, RepNZ, SG0) corresponds to VEX.nnn.f2

Having these additional matching clauses at the inner level
is a bit verbose, particularly since the "0x66" cases are in the
majority, but doing the decision tree in another way would
only be slightly shorter and probably harder to read.

* Fix sizes of memory operations in VPBROADCAST semantics

These have different sizes for destination and source and so
they need a bit more delicacy in decoding.

* Enforce VEX.LZ restriction in MULX decoding

As pointed out by Yan Peng, this ought to be checked, otherwise
we are quietly accepting a fake instruction that should cause
an undefined instruction exception on hardware.

* Add newly introduced AVX2 instructions to x86 allowed_asm list