SSE instructions needed for ML-KEM code (#293)
* Add MOVQ to x86 model

This tests the REX.W setting inside the existing decoding of MOVD.

* Add PCMPGTW to x86 model

* Add PSRLW to x86 model

And fix a few errors relating to MOVQ

* Add PSHUFB to x86 model

* Add PBLENDW to x86 model

* Add PINSRD and PINSRQ to x86 model

* Add PMOVMSKB to x86 model

The manual is a bit confusing about the destination size, and
assemblers differ in whether they allow 64-bit destination
registers. But the preferred disassembly seems to be 32-bit.
Since the result is just a zero-extension of a 16-bit value,
and 32-bit writes are then zero-extended to 64 bits anyway,
it makes no semantic difference.

* Fix core PMOVMSKB definition

* Add newly modelled SSE instructions to allowed_asm list

All appear in Intel's DOIT list so this should not be controversial:
https://www.intel.com/content/www/us/en/developer/articles/technical/software-security-guidance/resources/data-operand-independent-timing-instructions.html