Integer instructions needed for ML-KEM code (#292)
* Add POPCNT instruction to x86 model

* Add PEXT to x86 model

This is a somewhat exotic instruction, and our definition is very
clumsy and verbose.

* Enforce VEX.LZ restriction for newly added PEXT instruction

* Introduce word operation for the core x86 PEXT instruction

The previous PEXT definition resulted in a large and ugly expression.
This is now wrapped up via a general word operation word_condense
which is equipped with an evaluation conversion WORD_CONDENSE_CONV
inside the list extra_word_CONV.