Adopt enhanced WORD_SIMPLE_SUBWORD_CONV from latest HOL Light (#299)
This bumps the assumed HOL Light commit to update the inherited
conversion WORD_SIMPLE_SUBWORD_CONV. This allows some proofs to
be slightly simplified, although in the proofs of the ARM
Lenngren-style X25519, there is some additional compensation
needed because it now simplifies too much.

We also now take advantage of the fix to the type of usimd16 in HOL Light to
use it in PMOVMSKB (previously a usimd2+usimd8 workaround was used).