Change methodology for cosimulator's x86 instruction extraction
The previous version extracted the instructions from the formal
versions of the code inside the proof files. This has a few flaws,
notably the fact that sometimes line breaks separate the actual list
from the commented instruction, something that for example meant
that the important MULX instruction was *never* getting cosimulated.

The new version extracts directly from the source code by mangling +
assembly + objdump. This results in a strictly larger list of
instructions (2740 in place of 1342 currently). As well as the basic
"iclasses" list it also sets up a separate list "iclasses_simplemem"
of memory instructions, corresponding to memory instructions in the
code but with all relative addresses forced to be [rsp+32]. This
allows for at least a limited cosimulation of all the basic iforms
that occur in the code, except for control-flow and stack modification
(jcc, jmp, call, ret, push, pop).