Cosimulate simplified versions of x86 memory operations in the code
This supplements the existing memory operation cosimulation, which
has harnesses to handle general relative addresses. We now also
cosimulate "simplified" memory operations, which are basically
the memory iforms extracted from the code with all relative
addresses forced to be [rsp+32]. Collectively, this gives broader
coverage of memory operations for additional sanity checking. We
now select the kind of simulation with probability 1/3 each
when running the simulator:

 - Register-to-register operations
 - Memory operations with general relative address harness
 - Simplified memory operations extracted from the code