Cosimulate simplified versions of x86 memory operations in the code (#311)
* Force 32-alignment of RSP at start of x86 cosimulator template

This makes it easier to set up cosimulation of memory operations
like VMOVDQA that depend on this level of alignment. This wasn't
necessary before because 16-alignment is already the default on
many platforms.

* Change methodology for cosimulator's x86 instruction extraction

The previous version extracted the instructions from the formal
versions of the code inside the proof files. This has a few flaws,
notably the fact that sometimes line breaks separate the actual list
from the commented instruction, something that for example meant
that the important MULX instruction was *never* getting cosimulated.

The new version extracts directly from the source code by mangling +
assembly + objdump. This results in a strictly larger list of
instructions (2740 in place of 1342 currently). As well as the basic
"iclasses" list it also sets up a separate list "iclasses_simplemem"
of memory instructions, corresponding to memory instructions in the
code but with all relative addresses forced to be [rsp+32]. This
allows for at least a limited cosimulation of all the basic iforms
that occur in the code, except for control-flow and stack modification
(jcc, jmp, call, ret, push, pop).

* Cosimulate simplified versions of x86 memory operations in the code

This supplements the existing memory operation cosimulation, which
has harnesses to handle general relative addresses. We now also
cosimulate "simplified" memory operations, which are basically
the memory iforms extracted from the code with all relative
addresses forced to be [rsp+32]. Collectively, this gives broader
coverage of memory operations for additional sanity checking. We
now select the kind of simulation with probability 1/3 each
when running the simulator:

 - Register-to-register operations
 - Memory operations with general relative address harness
 - Simplified memory operations extracted from the code

* Add nonexecutable stack directive to x86 cosimulator template

This avoids pollution of the simulator output by linker warnings.

* Assume 32-alignment for x86 simple memory operation cosimulation

This is a bit over-conservative since it's really only needed for
VMOVDQA at the moment, but it keeps things simple.

* Fix issues and follow suggestions raised in reviews

These points were raised in reviews by @aqjune-aws and @pennyannn.

* Extend and comment the additional simple memory checks

* Clarify non-standard encodings in new x86 test cases

Be explicit that a couple of the new test cases are nonstandard
encodings, i.e. not what assemblers typically produce but still
valid in hardware. Also add one such more register case for MOVQ.