Add MOVSB to x86 model, with preliminary version of semantics
This supports the MOVSB instruction with or without the rep prefix;
the rep can be either canonical repz or repnz (both of which are
accepted by hardware). The underlying semantics is conservative,
only defining the result when the two regions don't overlap, but
to be accurate it should have the iterative definition in the
manuals. There is also currently zero support in the simulator.