x86 uniform rejection sampling for ML-KEM, MOVSB and related additions (#320)
This adds the x86 rejection sampling code from mlkem-native with a
couple of minor changes, and its correctness proof. The correctness
proof is in effect a simplified version of the ARM proof, with the
same kind of underlying logic but much simpler loop structure. The two
changes to the code are:

- Directly test for null input (buflen = 0), and skip the body and
  return 0 in that case.

- Use xmm3 in place of xmm6, so we don't need to save/restore it
  for Windows

The more complicated side of this update is the setup of
infrastructure to support a single instruction REP MOVSB, which is
used as the last step of the code to copy from the temporary buffer on
the stack to the output. This is rather different from all the x86
instructions formalized so far in that it manipulates
arbitrarily-sized buffers of memory, not just individual chunks 8, 64,
256 bits or whatever. Moreover the semantics is explicitly sequential,
and handling the case of overlapping input and output buffers needs to
take account of this. An auxiliary function x86_movsb is defined
recursively, and for simulation this is factored into an equivalent
using x86_stringcopy with bounded input and output buffers.

A second level of additional work is needed because repetition inside
REP MOVSB is controlled by the DF flag, not previously part of the x86
formalization (this determines whether iteration is downward or
upward). We add this to the model and the simulator, as well as the
two instructions for explicitly setting and clearing it (CLD, STD).
Finally, all the MOVSB and REP MOVSB semantics is checked in the
simulator for sizes up to 7 bytes, which should be enough to exercise
the various corner cases.